Attempt to change behaviour of 'loadResource':
It was initially planned to both offer a async and sync version of this function.

However, since this bundler is only meant to embed small (text) files, an asynchronous version is not really warranted.

Synchronous functions are also usable in async contexts where the opposite is not true.

Thus, making 'loadResource' synchronous makes the function more usable in more contexts.